ios : bonus

+++ app-ios-native/qingshow-ios-native/vc/user/setting/view/order/QSOrderListTableViewCell.h
@@ -64,5 +64,5 @@
 - (void)bindWithDict:(NSDictionary*)dict;
 
 @property (weak, nonatomic) NSObject<QSOrderListTableViewCellDelegate>* delegate;
-@property (assign,nonatomic) NSInteger type;
+@property (assign,nonatomic) int type;
 @end
